Preserve the directory structure when archiving modules in subdirectories
Instead of only putting the subdirectory contents into an
archive, let's preserve the directory structure in the archive
while still only including things from a particular
subdirectory. This way the archive has the same structure as the
cloned repository which makes it easier to work with.

Ticket: CFE-3780
Changelog: None